+++ packages/web-extension/src/walletManager/SignerManager/index.ts
@@ -1,6 +1,0 @@
-export * from './types';
-export * from './SignerManager';
-export * from './KeyAgentFactory';
-export * from './util';
-export * from './exposeSignerManagerApi';
-export * from './consumeSignerManagerApi';