+++ packages/web-extension/src/walletManager/SigningCoordinator/SigningCoordinator.ts
@@ -1,4 +1,3 @@
-/* eslint-disable brace-style */
 import { Cardano, Serialization } from '@cardano-sdk/core';
 import { CustomError } from 'ts-custom-error';
 import { InMemoryWallet, WalletType } from '../types';
@@ -12,19 +11,19 @@ import {
   SignOptions,
   SignRequest,
   SignTransactionProps,
-  SignerManagerConfirmationApi,
-  SignerManagerSignApi,
+  SigningCoordinatorConfirmationApi,
+  SigningCoordinatorSignApi,
   TransactionWitnessRequest
 } from './types';
 import { Subject } from 'rxjs';
 
 export type HardwareKeyAgentOptions = TrezorConfig;
 
-export type SignerManagerProps = {
+export type SigningCoordinatorProps = {
   hwOptions: HardwareKeyAgentOptions;
 };
 
-export type SignerManagerDependencies = {
+export type SigningCoordinatorDependencies = {
   keyAgentFactory: KeyAgentFactory;
 };
 
@@ -63,22 +62,24 @@ const bubbleResolveReject = async <R>(
   }
 };
 
-export class SignerManager<WalletMetadata extends {}>
-  implements SignerManagerConfirmationApi<WalletMetadata>, SignerManagerSignApi<WalletMetadata>
+export class SigningCoordinator<WalletMetadata extends {}, AccountMetadata extends {}>
+  implements
+    SigningCoordinatorConfirmationApi<WalletMetadata, AccountMetadata>,
+    SigningCoordinatorSignApi<WalletMetadata, AccountMetadata>
 {
-  readonly transactionWitnessRequest$ = new Subject<TransactionWitnessRequest<WalletMetadata>>();
-  readonly signDataRequest$ = new Subject<SignDataRequest<WalletMetadata>>();
+  readonly transactionWitnessRequest$ = new Subject<TransactionWitnessRequest<WalletMetadata, AccountMetadata>>();
+  readonly signDataRequest$ = new Subject<SignDataRequest<WalletMetadata, AccountMetadata>>();
   readonly #hwOptions: HardwareKeyAgentOptions;
   readonly #keyAgentFactory: KeyAgentFactory;
 
-  constructor(props: SignerManagerProps, { keyAgentFactory }: SignerManagerDependencies) {
+  constructor(props: SigningCoordinatorProps, { keyAgentFactory }: SigningCoordinatorDependencies) {
     this.#hwOptions = props.hwOptions;
     this.#keyAgentFactory = keyAgentFactory;
   }
 
   async signTransaction(
     { tx, signContext, options }: SignTransactionProps,
-    requestContext: RequestContext<WalletMetadata>
+    requestContext: RequestContext<WalletMetadata, AccountMetadata>
   ): Promise<Cardano.Signatures> {
     const transaction = Serialization.Transaction.fromCbor(tx);
     return this.#signRequest(
@@ -101,7 +102,10 @@ export class SignerManager<WalletMetadata extends {}>
     );
   }
 
-  async signData(props: SignDataProps, requestContext: RequestContext<WalletMetadata>): Promise<SignBlobResult> {
+  async signData(
+    props: SignDataProps,
+    requestContext: RequestContext<WalletMetadata, AccountMetadata>
+  ): Promise<SignBlobResult> {
     return this.#signRequest(
       this.signDataRequest$,
       {
@@ -113,7 +117,7 @@ export class SignerManager<WalletMetadata extends {}>
     );
   }
 
-  #signRequest<R, Req extends RequestBase<WalletMetadata> & SignRequest<R>>(
+  #signRequest<R, Req extends RequestBase<WalletMetadata, AccountMetadata> & SignRequest<R>>(
     emitter$: Subject<Req>,
     request: Omit<Req, 'reject' | 'sign'>,
     sign: (keyAgent: KeyAgent) => Promise<R>
@@ -141,7 +145,7 @@ export class SignerManager<WalletMetadata extends {}>
               sign: async (passphrase: Uint8Array, options?: SignOptions) =>
                 bubbleResolveReject(
                   async () => {
-                    const wallet = request.requestContext.wallet as InMemoryWallet<WalletMetadata>;
+                    const wallet = request.requestContext.wallet as InMemoryWallet<WalletMetadata, AccountMetadata>;
                     try {
                       const result = await sign(
                         this.#keyAgentFactory.InMemory({