+++ packages/web-extension/src/walletManager/SigningCoordinator/index.ts
@@ -1,0 +1,6 @@
+export * from './types';
+export * from './SigningCoordinator';
+export * from './KeyAgentFactory';
+export * from './util';
+export * from './exposeSigningCoordinatorApi';
+export * from './consumeSigningCoordinatorApi';