+++ packages/web-extension/src/walletManager/WalletRepository/WalletRepository.ts
@@ -1,5 +1,12 @@
-import { AddAccountProps, AddWalletProps, RemoveAccountProps, UpdateMetadataProps, WalletRepositoryApi } from './types';
-import { AnyWallet, ScriptWallet, WalletId, WalletType } from '../types';
+import {
+  AddAccountProps,
+  AddWalletProps,
+  RemoveAccountProps,
+  UpdateAccountMetadataProps,
+  UpdateWalletMetadataProps,
+  WalletRepositoryApi
+} from './types';
+import { AnyWallet, WalletId, WalletType } from '../types';
 import { Logger } from 'ts-log';
 import { Observable, defer, firstValueFrom, map, shareReplay, switchMap, take } from 'rxjs';
 import { WalletConflictError } from '../errors';
@@ -7,8 +14,8 @@ import { contextLogger } from '@cardano-sdk/util';
 import { getWalletId } from '../util';
 import { storage } from '@cardano-sdk/wallet';
 
-export interface WalletRepositoryDependencies<AccountMetadata extends {}> {
-  store: storage.CollectionStore<AnyWallet<AccountMetadata>>;
+export interface WalletRepositoryDependencies<WalletMetadata extends {}, AccountMetadata extends {}> {
+  store: storage.CollectionStore<AnyWallet<WalletMetadata, AccountMetadata>>;
   logger: Logger;
 }
 
@@ -18,8 +25,8 @@ const cloneSplice = <T>(array: T[], start: number, deleteCount: number, ...items
   ...array.slice(start + deleteCount)
 ];
 
-const findAccount = <AccountMetadata extends {}>(
-  wallets: AnyWallet<AccountMetadata>[],
+const findAccount = <WalletMetadata extends {}, AccountMetadata extends {}>(
+  wallets: AnyWallet<WalletMetadata, AccountMetadata>[],
   walletId: WalletId,
   accountIndex: number
 ) => {
@@ -36,12 +43,14 @@ const findAccount = <AccountMetadata extends {}>(
   };
 };
 
-export class WalletRepository<AccountMetadata extends {}> implements WalletRepositoryApi<AccountMetadata> {
+export class WalletRepository<WalletMetadata extends {}, AccountMetadata extends {}>
+  implements WalletRepositoryApi<WalletMetadata, AccountMetadata>
+{
   readonly #logger: Logger;
-  readonly #store: WalletRepositoryDependencies<AccountMetadata>['store'];
-  readonly wallets$: Observable<AnyWallet<AccountMetadata>[]>;
+  readonly #store: WalletRepositoryDependencies<WalletMetadata, AccountMetadata>['store'];
+  readonly wallets$: Observable<AnyWallet<WalletMetadata, AccountMetadata>[]>;
 
-  constructor({ logger, store }: WalletRepositoryDependencies<AccountMetadata>) {
+  constructor({ logger, store }: WalletRepositoryDependencies<WalletMetadata, AccountMetadata>) {
     this.#store = store;
     this.#logger = contextLogger(logger, 'WalletRepository');
     this.wallets$ = defer(() => store.observeAll()).pipe(shareReplay(1));
@@ -55,7 +64,7 @@ export class WalletRepository<AccountMetadata extends {}> implements WalletRepos
     );
   }
 
-  async addWallet(props: AddWalletProps<AccountMetadata>): Promise<WalletId> {
+  async addWallet(props: AddWalletProps<WalletMetadata, AccountMetadata>): Promise<WalletId> {
     this.#logger.debug('addWallet', props.type);
     const walletId = await getWalletId(
       props.type === WalletType.Script ? props.script : props.extendedAccountPublicKey
@@ -129,39 +138,53 @@ export class WalletRepository<AccountMetadata extends {}> implements WalletRepos
     );
   }
 
-  updateMetadata(props: UpdateMetadataProps<AccountMetadata>): Promise<UpdateMetadataProps<AccountMetadata>> {
-    const { walletId, accountIndex, metadata } = props;
-    this.#logger.debug('updateMetadata', walletId, accountIndex, metadata);
+  updateWalletMetadata(
+    props: UpdateWalletMetadataProps<WalletMetadata>
+  ): Promise<UpdateWalletMetadataProps<WalletMetadata>> {
+    const { walletId, metadata } = props;
+    this.#logger.debug('updateWalletMetadata', walletId, metadata);
     return firstValueFrom(
       this.#getWallets().pipe(
         switchMap((wallets) => {
-          if (typeof accountIndex !== 'undefined') {
-            const bip32Account = findAccount(wallets, walletId, accountIndex);
-            if (!bip32Account) {
-              throw new WalletConflictError(`Account not found: ${walletId}/${accountIndex}`);
-            }
+          const walletIndex = wallets.findIndex((wallet) => wallet.walletId === walletId);
+          if (walletIndex >= 0) {
+            // update any wallet
             return this.#store.setAll(
-              cloneSplice(wallets, bip32Account.walletIdx, 1, {
-                ...bip32Account.wallet,
-                accounts: cloneSplice(bip32Account.wallet.accounts, bip32Account.accountIdx, 1, {
-                  ...bip32Account.account,
-                  metadata
-                })
-              })
-            );
-          }
-          const scriptWalletIndex = wallets.findIndex(
-            (wallet) => wallet.walletId === walletId && wallet.type === WalletType.Script
-          );
-          if (scriptWalletIndex >= 0) {
-            return this.#store.setAll(
-              cloneSplice(wallets, scriptWalletIndex, 1, {
-                ...(wallets[scriptWalletIndex] as ScriptWallet<AccountMetadata>),
+              cloneSplice(wallets, walletIndex, 1, {
+                ...(wallets[walletIndex] as AnyWallet<WalletMetadata, AccountMetadata>),
                 metadata
               })
             );
           }
-          throw new WalletConflictError(`Script wallet not found: ${walletId}`);
+          throw new WalletConflictError(`Wallet not found: ${walletId}`);
+        }),
+        map(() => props)
+      )
+    );
+  }
+
+  updateAccountMetadata(
+    props: UpdateAccountMetadataProps<AccountMetadata>
+  ): Promise<UpdateAccountMetadataProps<AccountMetadata>> {
+    const { walletId, accountIndex, metadata } = props;
+    this.#logger.debug('updateAccountMetadata', walletId, accountIndex, metadata);
+    return firstValueFrom(
+      this.#getWallets().pipe(
+        switchMap((wallets) => {
+          // update account
+          const bip32Account = findAccount(wallets, walletId, accountIndex);
+          if (!bip32Account) {
+            throw new WalletConflictError(`Account not found: ${walletId}/${accountIndex}`);
+          }
+          return this.#store.setAll(
+            cloneSplice(wallets, bip32Account.walletIdx, 1, {
+              ...bip32Account.wallet,
+              accounts: cloneSplice(bip32Account.wallet.accounts, bip32Account.accountIdx, 1, {
+                ...bip32Account.account,
+                metadata
+              })
+            })
+          );
         }),
         map(() => props)
       )