+++ packages/hardware-trezor/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.6](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/hardware-trezor@0.4.5...@cardano-sdk/hardware-trezor@0.4.6) (2024-01-17)
+
+**Note:** Version bump only for package @cardano-sdk/hardware-trezor
+
 ## [0.4.5](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/hardware-trezor@0.4.4...@cardano-sdk/hardware-trezor@0.4.5) (2024-01-05)
 
 ### Features