+++ packages/input-selection/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.13](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/input-selection@0.12.12...@cardano-sdk/input-selection@0.12.13) (2024-01-17)
+
+**Note:** Version bump only for package @cardano-sdk/input-selection
+
 ## [0.12.12](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/input-selection@0.12.11...@cardano-sdk/input-selection@0.12.12) (2024-01-05)
 
 **Note:** Version bump only for package @cardano-sdk/input-selection