+++ packages/ogmios/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.15.8](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/ogmios@0.15.7...@cardano-sdk/ogmios@0.15.8) (2024-01-17)
+
+**Note:** Version bump only for package @cardano-sdk/ogmios
+
 ## [0.15.7](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/ogmios@0.15.6...@cardano-sdk/ogmios@0.15.7) (2024-01-05)
 
 **Note:** Version bump only for package @cardano-sdk/ogmios